+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.12-cloud] - 2023-11-05
 ### Changed
 - support '/' for workload name in workload wizard
+- bad request error for duplicate variable in batch fetch secret
 
 ## [1.0.11-cloud] - 2023-10-29
 ### Changed